# backend/docs/swagger.yaml
basePath: /
definitions:
  activity.Activity:
    properties:
      activity_image:
        type: string
      creator_id:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
...
  responses.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  responses.UserCreateRequest:
    properties:
      email:
        example: user@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: password123
        type: string
    type: object
  user.User:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for CODECK - COmpetição DE Código do Kim. A platform for competitive
    programmers.
  license:
    name: MIT
    url: https://github.com/your-username/codeck/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: CODECK API
  version: "1.0"
paths:
...
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, name, and password
      parameters:
      - description: User creation data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/responses.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get user information by user ID (password field excluded)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get user by ID
      tags:
      - users
  /users/{id}/activities:
    get:
      consumes:
      - application/json
      description: Get all activities created by a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/activity.Activity'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get user activities
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"